Use the admin hooks
