+++ src/hooks/useMentor.ts
@@ -1,11 +1,9 @@
-import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
+import { useQuery } from '@tanstack/react-query';
 import { API_URL } from '../constants';
 import axios from 'axios';
 import { type Mentor } from '../types';
 
 const useMentor = (mentorId: string | undefined) => {
-  const queryClient = useQueryClient();
-
   const { isLoading, error, data } = useQuery({
     queryKey: ['mentor', mentorId],
     initialData: null,
@@ -18,28 +16,10 @@ const useMentor = (mentorId: string | undefined) => {
     },
   });
 
-  const { mutate: changeState } = useMutation({
-    mutationFn: async (newState: string) => {
-      if (mentorId != null) {
-        const { data } = await axios.put(
-          `${API_URL}/admin/mentors/${mentorId}/state`,
-          {
-            state: newState,
-          }
-        );
-        return data;
-      }
-    },
-    onSuccess: () => {
-      void queryClient.invalidateQueries({ queryKey: ['mentor', mentorId] });
-    },
-  });
-
   return {
     isLoading,
     error,
     data,
-    changeState,
   };
 };
